+++ lib/assets/javascripts/new-dashboard/pages/Data/Catalog.vue
@@ -1,36 +1,36 @@
 <template>
-  <section class="catalogue">
+  <section class="catalog">
     <div class="container grid">
       <div class="full-width">
         <SectionTitle class="grid-cell">
           <template slot="icon">
-            <img src="../../assets/icons/section-title/catalogue.svg" width="18" height="20" />
+            <img src="../../assets/icons/section-title/catalog.svg" width="18" height="20" />
           </template>
           <template slot="title">
-              <span>{{ $t('CataloguePage.header') }}</span>
+              <span>{{ $t('CatalogPage.header') }}</span>
           </template>
         </SectionTitle>
       </div>
     </div>
 
-    <div class="catalogue__filter container grid" action="">
-      <div class="grid-cell grid-cell--col6 catalogue__filter--dropdown">
-        <CatalogueDropdown ref="dropdownCategories"
-          :title="$t('CatalogueDropdown.category.title')"
-          :placeholder="$t('CatalogueDropdown.category.placeholder')"
+    <div class="catalog__filter container grid" action="">
+      <div class="grid-cell grid-cell--col6 catalog__filter--dropdown">
+        <CatalogDropdown ref="dropdownCategories"
+          :title="$t('CatalogDropdown.category.title')"
+          :placeholder="$t('CatalogDropdown.category.placeholder')"
           :options="categories"
           :disabled="true"
           @selected="getCountries"
           @reset="resetCategory">
           <template slot="extra">
-            <span v-html="$t('CatalogueDropdown.category.extra')" class="text is-small"></span>
+            <span v-html="$t('CatalogDropdown.category.extra')" class="text is-small"></span>
           </template>
-        </CatalogueDropdown>
+        </CatalogDropdown>
       </div>
-      <div class="grid-cell grid-cell--col6 catalogue__filter--dropdown">
-        <CatalogueDropdown ref="dropdownCountries"
-          :title="$t('CatalogueDropdown.country.title')"
-          :placeholder="$t('CatalogueDropdown.country.placeholder')"
+      <div class="grid-cell grid-cell--col6 catalog__filter--dropdown">
+        <CatalogDropdown ref="dropdownCountries"
+          :title="$t('CatalogDropdown.country.title')"
+          :placeholder="$t('CatalogDropdown.country.placeholder')"
           :options="countries"
           :disabled="true"
           :limitHeight="true"
@@ -38,25 +38,25 @@
           @selected="getDatasets"
           @reset="resetCountry">
           <template slot="extra">
-            <span v-html="$t('CatalogueDropdown.country.extra')" class="text is-small"></span>
+            <span v-html="$t('CatalogDropdown.country.extra')" class="text is-small"></span>
           </template>
-        </CatalogueDropdown>
+        </CatalogDropdown>
       </div>
     </div>
 
     <!-- Datasets list -->
     <div class="container grid"  v-if="datasets.length > 0">
       <div class="grid-cell grid-cell--col12">
-        <CatalogueListHeader :order="appliedOrder" :orderDirection="appliedOrderDirection" @changeOrder="applyOrder"></CatalogueListHeader>
+        <CatalogListHeader :order="appliedOrder" :orderDirection="appliedOrderDirection" @changeOrder="applyOrder"></CatalogListHeader>
         <ul v-if="!isFetchingDatasets" >
           <li v-for="(dataset) in currentPageDatasets" :key="dataset.id">
-            <CatalogueCard :dataset="dataset"></CatalogueCard>
+            <CatalogCard :dataset="dataset"></CatalogCard>
           </li>
         </ul>
 
         <ul v-else>
           <li v-for="(n) in resultsPerPage" :key="n">
-            <CatalogueFakeCard />
+            <CatalogFakeCard />
           </li>
         </ul>
       </div>
@@ -68,28 +68,28 @@
 import { mapState } from 'vuex';
 import sendCustomEvent from 'new-dashboard/utils/send-custom-event';
 import SectionTitle from 'new-dashboard/components/SectionTitle';
-import CatalogueListHeader from 'new-dashboard/components/Catalogue/CatalogueListHeader';
-import CatalogueCard from 'new-dashboard/components/Catalogue/CatalogueCard';
-import CatalogueFakeCard from 'new-dashboard/components/Catalogue/CatalogueFakeCard';
-import CatalogueDropdown from 'new-dashboard/components/Catalogue/CatalogueDropdown';
+import CatalogListHeader from 'new-dashboard/components/Catalog/CatalogListHeader';
+import CatalogCard from 'new-dashboard/components/Catalog/CatalogCard';
+import CatalogFakeCard from 'new-dashboard/components/Catalog/CatalogFakeCard';
+import CatalogDropdown from 'new-dashboard/components/Catalog/CatalogDropdown';
 import Pagination from 'new-dashboard/components/Pagination';
 
 export default {
-  name: 'CataloguePage',
+  name: 'CatalogPage',
   components: {
     SectionTitle,
-    CatalogueListHeader,
-    CatalogueCard,
-    CatalogueFakeCard,
-    CatalogueDropdown,
+    CatalogListHeader,
+    CatalogCard,
+    CatalogFakeCard,
+    CatalogDropdown,
     Pagination
   },
   beforeRouteUpdate (to, from, next) {
-    sendCustomEvent('catalogueSelectCategory', {
-      catalogueSelectedCategory: to.query.category
+    sendCustomEvent('catalogSelectCategory', {
+      catalogSelectedCategory: to.query.category
     });
-    sendCustomEvent('catalogueSelectCountry', {
-      catalogueSelectedCountry: to.query.country
+    sendCustomEvent('catalogSelectCountry', {
+      catalogSelectedCountry: to.query.country
     });
     next();
   },
@@ -104,16 +104,16 @@ export default {
   },
   computed: {
     ...mapState({
-      numPages: state => state.catalogue.numPages,
-      currentPage: state => state.catalogue.page,
-      datasets: state => state.catalogue.list,
-      categories: state => state.catalogue.categories,
-      countries: state => state.catalogue.countries,
-      isFetchingDatasets: state => state.catalogue.isFetching,
-      numResults: state => state.catalogue.umResults,
-      resultsPerPage: state => state.catalogue.resultsPerPage,
-      appliedOrder: state => state.catalogue.order,
-      appliedOrderDirection: state => state.catalogue.orderDirection
+      numPages: state => state.catalog.numPages,
+      currentPage: state => state.catalog.page,
+      datasets: state => state.catalog.list,
+      categories: state => state.catalog.categories,
+      countries: state => state.catalog.countries,
+      isFetchingDatasets: state => state.catalog.isFetching,
+      numResults: state => state.catalog.umResults,
+      resultsPerPage: state => state.catalog.resultsPerPage,
+      appliedOrder: state => state.catalog.order,
+      appliedOrderDirection: state => state.catalog.orderDirection
     }),
     shouldShowPagination () {
       return !this.isFetchingDatasets && this.numPages > 1;
@@ -126,14 +126,14 @@ export default {
     goToPage (page) {
       window.scroll({ top: 0, left: 0 });
       this.$router.push({
-        name: 'catalogue',
+        name: 'catalog',
         params: this.$route.params,
         query: { ...this.$route.query, page }
       });
     },
     applyOrder (orderParams) {
       this.$router.push({
-        name: 'catalogue',
+        name: 'catalog',
         params: this.$route.params,
         query: {
           ...this.$route.query,
@@ -142,11 +142,11 @@ export default {
           order_direction: orderParams.direction
         }
       });
-      this.$store.dispatch('catalogue/order', orderParams);
+      this.$store.dispatch('catalog/order', orderParams);
     },
     getCategories () {
       this.clearList();
-      this.$store.dispatch('catalogue/fetchCategories')
+      this.$store.dispatch('catalog/fetchCategories')
         .then(
           () => {
             this.$refs.dropdownCategories.enableDropdown();
@@ -160,7 +160,7 @@ export default {
     },
     getCountries (category) {
       this.clearList();
-      this.$store.dispatch('catalogue/fetchCountries', category)
+      this.$store.dispatch('catalog/fetchCountries', category)
         .then(
           () => {
             this.$refs.dropdownCountries.clearInput();
@@ -174,13 +174,13 @@ export default {
         );
     },
     getDatasets (country) {
-      this.$store.dispatch('catalogue/fetchDatasets', {
+      this.$store.dispatch('catalog/fetchDatasets', {
         category: this.$refs.dropdownCategories.selected,
         country
       }).then(
         () => {
           this.$router.push({
-            name: 'catalogue',
+            name: 'catalog',
             query: {
               category: this.$refs.dropdownCategories.selected,
               country
@@ -203,17 +203,17 @@ export default {
       this.clearList();
     },
     clearList () {
-      this.$store.dispatch('catalogue/clearList');
+      this.$store.dispatch('catalog/clearList');
     },
     setUrlOptions () {
       const category = this.$route.query.category;
       const country = this.$route.query.country;
-      const promiseDatasets = this.$store.dispatch('catalogue/fetchDatasets', {
+      const promiseDatasets = this.$store.dispatch('catalog/fetchDatasets', {
         category: category,
         country: country
       });
-      const promiseCategories = this.$store.dispatch('catalogue/fetchCategories');
-      const promiseCountires = this.$store.dispatch('catalogue/fetchCountries', category);
+      const promiseCategories = this.$store.dispatch('catalog/fetchCategories');
+      const promiseCountires = this.$store.dispatch('catalog/fetchCountries', category);
       Promise.all([promiseDatasets, promiseCategories, promiseCountires])
         .then(
           () => {
@@ -222,11 +222,11 @@ export default {
               this.$refs.dropdownCountries.setInput(country);
             } else {
               this.getCategories();
-              this.$router.push({ name: 'catalogue' });
+              this.$router.push({ name: 'catalog' });
             }
           },
           () => {
-            this.$router.push({ name: 'catalogue' });
+            this.$router.push({ name: 'catalog' });
           }
         );
     }
@@ -237,7 +237,7 @@ export default {
 <style scoped lang="scss">
 @import 'new-dashboard/styles/variables';
 
-.catalogue {
+.catalog {
   min-height: 640px;
   margin-top: 64px;